Bug Fixes to Jumping Platforms & CameraController
New:
- Updated Instakill Spikes Sprites

Changes:
- JumpPlatform script used to create platforms that bounces the player
continuously if they jump on it is working.
- CameraController script fixed, so that the camera will now correctly
follow the player.
- Adjusted the Level Bounds Object in Unity.
- Created animations for Instakill Spikes

Bugs:
- The checkpoints are still buggy. Player will start at the starting
checkpoint. If killed by the first Instakill Spikes, the player will
respawn at the first checkpoint. If the player is killed on the second
Instakill Spikes, the player will respawn at the starting checkpoint.
The Managers Game Object sets the Debug Spawn to the Level Start
Checkpoint.

TODO:
- Fix checkpoints, so that player correctly spawns.