Project Cleanup & Minor Bug Fixes
New:
- none

Updates:
- all scripts with RespawnListener implemented no longer re-initializes
the GameObject at _startposition or RespawnPosition. This was causing
the spawning of GameObjects located at (0,0)
- Player script comments, swimming had been removed.
- Player script hInput, and vInput set in update() function instead of
the HandleInput() function.

Bugs:
- none aware

Note:
- EndLevelMenu takes precedence over the time of the game. Even if the
player clicks pause during game, it will not set time to 0f. Possible
fix is to create an end level screen (for win), and a lose screen as
separate screens.